wip: MenuCRUD実装途中

+++ database/migrations/2026_05_29_000004_create_menus_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('menus', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->unsignedTinyInteger('dish_category');
+            $table->integer('dish_category');
             $table->integer('calorie')->nullable();
             $table->timestamps();
         });
